+++ metadata/PgPlaces.js
@@ -12,16 +12,20 @@ class PgPlaces extends PgCollection {
      * @param schema {String} Schema containing data tables.
      */
     constructor(pool, schema) {
-        super(pool, schema, PgPlaces.tableName());
+        super(pool, schema, `PgPlaces`);
     }
 
     /**
      * @inheritDoc
      */
-    create(id, userId) {
-        return super.create(id, userId).then(() => {
-            return this._pool.query(`INSERT INTO ${this._schema}.${PgPlaces.tableName()} (id) VALUES (${id})`);
-        });
+    create(id) {
+        if(!id) {
+            throw new Error(
+                logger.error(`${this._name}#create Id must be provided.`)
+            );
+        }
+
+        return this._pool.query(`INSERT INTO ${this._schema}.${PgPlaces.tableName()} (id) VALUES (${id});`);
     }
 
     /**
@@ -32,43 +36,51 @@ class PgPlaces extends PgCollection {
      * @param place.name {String} Name of the place
      * @param place.bbox {String} Bbox associated with given place.
      * @param place.dataset {String} Scope of the place
-     * @param userId {Number} Id of the user updating place
      */
-    update(id, place, userId) {
-        return super.update(id, place, userId).then(() => {
-            let sql = `BEGIN TRANSACTION; `;
-            let changes = [];
+    update(id, place) {
+        if(!id) {
+            throw new Error(
+                logger.error(`${this._name}#update Id must be provided.`)
+            );
+        }
+        if(!place) {
+            throw new Error(
+                logger.error(`${this._name}#update Updated object must be provided.`)
+            );
+        }
 
-            if(place.name !== null) {
-                changes.push(` name = '${name}' `);
-            }
-            if(place.bbox !== null) {
-                changes.push(` bbox = ${place.bbox} `);
-            }
-            if(place.dataset !== null) {
-                changes.push(` dataset = ${place.dataset} `);
-            }
+        let sql = ``;
+        let changes = [];
 
-            if(changes.length > 0) {
-                sql += `UPDATE ${this._schema}.${PgPlaces.tableName()} SET ${changes.join(',')} WHERE id = ${id}; `;
-            }
-            sql += `COMMIT; `;
+        if(place.name !== null) {
+            changes.push(` name = '${name}' `);
+        }
+        if(place.bbox !== null) {
+            changes.push(` bbox = ${place.bbox} `);
+        }
+        if(place.dataset !== null) {
+            changes.push(` dataset = ${place.dataset} `);
+        }
 
-            logger.info(`PgPlaces#update SQL: ${sql}`);
-            return this._pool.query(sql);
-        }).catch(err => {
-            logger.error(`PgPlaces#update ERROR: `, err);
-            return this._pool.query(`ROLLBACK`);
-        });
+        if(changes.length > 0) {
+            sql += `UPDATE ${this._schema}.${PgPlaces.tableName()} SET ${changes.join(',')} WHERE id = ${id}; `;
+        }
+
+        logger.info(`PgPlaces#update SQL: ${sql}`);
+        return this._pool.query(sql);
     }
 
     /**
      * @inheritDoc
      */
-    delete(id, userId) {
-        return super.delete(id, userId).then(() => {
-            return this._pool.query(`DELETE FROM ${this._schema}.${PgPlaces.tableName()} WHERE id = ${id} CASCADE`);
-        });
+    delete(id) {
+        if(!id) {
+            throw new Error(
+                logger.error(`${this._name}#delete Id must be provided.`)
+            );
+        }
+
+        return this._pool.query(`DELETE FROM ${this._schema}.${PgPlaces.tableName()} WHERE id = ${id} CASCADE`);
     }
 
     static tableName() {
